# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certif",
        "description": "Api de l'appli PENC",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/displayreply": {
            "get": {
                "tags": [
                    "Replies"
                ],
                "summary": "Récupère tous les messages et leurs réponses",
                "description": "Récupère tous les messages et leurs réponses.",
                "operationId": "b06c7812fc8b25d84806a5e7ae0879d5",
                "responses": {
                    "200": {
                        "description": "Liste de messages et leurs réponses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Reply"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Une erreur inattendue est survenue"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sendreply": {
            "post": {
                "tags": [
                    "Replies"
                ],
                "summary": "Crée une nouvelle réponse à un message",
                "description": "Crée une nouvelle réponse à un message.",
                "operationId": "b620f1a3af2e91e887011b33c34bde91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reply_content",
                                    "message_id"
                                ],
                                "properties": {
                                    "reply_content": {
                                        "type": "string",
                                        "example": "Contenu de la réponse"
                                    },
                                    "message_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre reponse a bien été créé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Un administrateur ne peut pas envoyer de reponse."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de la création de la reponse."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updatespecificreply/{reply}": {
            "put": {
                "tags": [
                    "Replies"
                ],
                "summary": "Met à jour une réponse existante",
                "description": "Met à jour une réponse existante.",
                "operationId": "a2a2e6c709b55c6f6fd54a0a11fdf62c",
                "parameters": [
                    {
                        "name": "reply",
                        "in": "path",
                        "description": "ID de la réponse à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reply_content",
                                    "user_id",
                                    "message_id"
                                ],
                                "properties": {
                                    "reply_content": {
                                        "type": "string",
                                        "example": "Nouveau contenu de la réponse"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifications effectuées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Modifications effectuées avec succès."
                                        },
                                        "resultat": {
                                            "$ref": "#/components/schemas/Reply"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le contenu est requis."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne. Veuillez réessayer."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Message d'erreur spécifique"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/deletespecificreply/{reply}": {
            "delete": {
                "tags": [
                    "Replies"
                ],
                "summary": "Supprime une réponse existante",
                "description": "Supprime une réponse existante.",
                "operationId": "51bad72b96fdb3fa6073ea9168dfee0a",
                "parameters": [
                    {
                        "name": "reply",
                        "in": "path",
                        "description": "ID de la réponse à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le réponse a été supprimé avec succès par l'administrateur."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'êtes pas autorisé à supprimer cette réponse."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}